requirements.txt
torch>=2.1.0
transformers>=4.40.0,<4.56.0
pillow>=10.0.0
accelerate>=0.27.0
requests>=2.31.0
huggingface_hub>=0.20.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

einops>=0.7.0
timm>=0.9.0
